# openapi.yaml
openapi: 3.1.0
info:
  title: Wiki JA Recommender API
  version: "1.0.0"
servers:
  - url: https://wikichat.urayaha.tokyo 
paths:
  /health:
    get:
      operationId: health
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: object
                properties:
                  ok:
                    type: boolean
  /next_article:
    get:
      operationId: nextArticle
      parameters:
        - in: query
          name: user
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Next unseen article for the user
          content:
            application/json:
              schema:
                type: object
                properties:
                  article_id: { type: integer }
                  title: { type: string }
                  url: { type: string }
                  summary:
                    type: object
                    properties:
                      extract: { type: string }
                      thumbnail: { type: string }
        "404":
          description: No unseen article found
  /react:
    post:
      operationId: react
      parameters:
        - in: query
          name: user
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [article_id, reaction]
              properties:
                article_id:
                  type: integer
                reaction:
                  type: string
                  enum: [like, skip, block]
      responses:
        "200":
          description: ok
  /article_content:
    get:
      operationId: articleContent
      parameters:
        - in: query
          name: url
          required: true
          schema:
            type: string
        - in: query
          name: format
          required: false
          schema:
            type: string
            enum: [markdown, text]
            default: markdown
      responses:
        "200":
          description: Normalized article content
          content:
            application/json:
              schema:
                type: object
                properties:
                  title: { type: string, nullable: true }
                  url: { type: string }
                  format: { type: string, enum: [markdown, text] }
                  content: { type: string }